Move is_gas_exceeded to where it's used
